+++ inspector/src/components/actionTabs/lines/vector3LineComponent.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Vector3 } from "babylonjs/Maths/math.vector";
 import { Observable } from "babylonjs/Misc/observable";
 
-import { NumericInputComponent } from "./numericInputComponent";
+import { NumericInputComponent } from "./../../../sharedUiComponents/lines/numericInputComponent";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faMinus, faPlus } from "@fortawesome/free-solid-svg-icons";
 import { PropertyChangedEvent } from "../../propertyChangedEvent";
