+++ inspector/src/components/actionTabs/tabs/propertyGrids/animations/editorControls.tsx
@@ -3,7 +3,7 @@ import { Observable } from "babylonjs/Misc/observable";
 import { PropertyChangedEvent } from "../../../../../components/propertyChangedEvent";
 import { Animation } from "babylonjs/Animations/animation";
 import { IconButtonLineComponent } from "../../../lines/iconButtonLineComponent";
-import { NumericInputComponent } from "../../../lines/numericInputComponent";
+import { NumericInputComponent } from "../../../../../sharedUiComponents/lines/numericInputComponent";
 import { AddAnimation } from "./addAnimation";
 import { AnimationListTree, SelectedCoordinate } from "./animationListTree";
 import { IAnimatable } from "babylonjs/Animations/animatable.interface";
