+++ sharedUiComponents/lines/messageLineComponent.tsx
