+++ sharedUiComponents/lines/numericInputComponent.tsx
